requirements/base.txt
aioredis==1.3.0
amqp==2.5.1
appdirs==1.4.4
appnope==0.1.0
asgiref==3.2.2
asn1crypto==0.24.0
astroid==2.4.1
async-timeout==3.0.1
attrs==19.1.0
autobahn==19.10.1
Automat==0.7.0
backcall==0.1.0
billiard==3.6.1.0
certifi==2020.4.5.2
cffi==1.12.3
channels==2.3.0
channels-redis==2.4.0
chardet==3.0.4
click==7.1.2
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.7
daphne==2.3.0
decorator==4.4.0
Django==2.2.5
django-cors-headers==3.1.0
django-environ==0.4.5
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
entrypoints==0.3
hiredis==1.0.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==0.23
incremental==17.5.0
ipython==7.8.0
ipython-genutils==0.2.0
isort==4.3.21
itypes==1.2.0
jedi==0.15.1
Jinja2==2.11.2
kombu==4.6.5
lazy-object-proxy==1.4.3
Markdown==3.2.2
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==7.2.0
msgpack==0.6.2
parso==0.5.1
pathspec==0.8.0
pexpect==4.7.0
pickleshare==0.7.5
prompt-toolkit==2.0.9
psycopg2==2.8.3
ptyprocess==0.6.0
pycodestyle==2.5.0
pycparser==2.19
pyflakes==2.1.1
Pygments==2.4.2
PyHamcrest==1.9.0
PyJWT==1.7.1
pytz==2019.2
PyYAML==5.1.2
redis==3.3.11
regex==2020.5.14
requests==2.24.0
six==1.12.0
sqlparse==0.3.0
toml==0.10.1
traitlets==4.3.2
Twisted==19.7.0
txaio==18.8.1
typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.25.9
vine==1.3.0
wcwidth==0.1.7
wrapt==1.12.1
zipp==0.6.0
zope.interface==4.6.0
django-background-tasks==1.2.5
django-compat==1.0.15
django-polymorphic==2.1.2
django-rest-polymorphic==0.1.9
argparse==1.4.0
dateutils==0.6.8
python-dateutil==2.8.1
aiohttp==3.6.2
multidict==4.7.6
yarl==1.4.2
numpy==1.19.1
scipy==1.5.2
investpy==1.0